backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # Optional - only needed for PostgreSQL. For SQLite, skip this.
# langgraph==0.0.20  # Optional - using custom orchestrator
# langchain==0.1.0  # Optional
# langchain-google-genai==0.0.6  # Optional
networkx==3.2.1
gitpython==3.1.40
tree-sitter==0.20.4
tree-sitter-python==0.25.0
tree-sitter-javascript==0.25.0
tree-sitter-typescript==0.23.2
aiofiles==23.2.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

